# apps/ios/Packages/DenAPI/Sources/DenAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Den API",
    "description": "M3 REST API. Cookie-authenticated writes require Origin and X-CSRF-Token. CLI and agents use bearer authentication. Entity IDs are ULIDs; background image IDs are opaque SHA-256 content IDs.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/channels/{id}/read": {
      "put": {
        "tags": [
          "inbox"
        ],
        "operationId": "put__channels__id__read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkRead"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelReadState"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/me/notification-preferences": {
      "get": {
        "tags": [
          "inbox"
        ],
        "operationId": "get__users_me_notification_preferences",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "inbox"
        ],
        "operationId": "put__users_me_notification_preferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPreferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/me/read-state": {
      "get": {
        "tags": [
          "inbox"
        ],
        "operationId": "get__users_me_read_state",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelReadState"
                  }
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ChannelReadState": {
        "type": "object",
        "required": [
          "channel_id",
          "unread_count",
          "mention_count",
          "notification_count"
        ],
        "properties": {
          "channel_id": {
            "$ref": "#/components/schemas/String"
          },
          "last_read_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "mention_count": {
            "type": "integer",
            "format": "int64"
          },
          "notification_count": {
            "type": "integer",
            "format": "int64",
            "description": "Unread messages eligible under current preferences, excluding own messages."
          },
          "unread_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "MarkRead": {
        "type": "object",
        "required": [
          "message_id"
        ],
        "properties": {
          "message_id": {
            "$ref": "#/components/schemas/String"
          },
          "roots_only": {
            "type": "boolean",
            "description": "Legacy false marks the flat conversation through this message. New clients\nsend true for the main timeline and read threads at the thread endpoint."
          }
        }
      },
...
      "NotificationPreferences": {
        "type": "object",
        "required": [
          "mentions",
          "dms",
          "subscribed_channel_ids"
        ],
        "properties": {
          "dms": {
            "type": "boolean"
          },
          "mentions": {
            "type": "boolean"
          },
          "subscribed_channel_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          }
        }
      },
...
      "String": {
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "den_session"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ]
}